@@ -14,16 +14,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.0.3",
+    "@angular/animations": "^14.3.0",
     "@angular/cdk": "^13.1.0",
-    "@angular/common": "~13.0.3",
-    "@angular/compiler": "~13.0.3",
-    "@angular/core": "~13.0.3",
-    "@angular/forms": "~13.0.3",
-    "@angular/platform-browser": "~13.0.3",
-    "@angular/platform-browser-dynamic": "~13.0.3",
-    "@angular/router": "~13.0.3",
-    "@angular/service-worker": "~13.0.3",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "@angular/service-worker": "^14.3.0",
     "@ngrx/effects": "^13.0.2",
     "@ngrx/router-store": "^13.0.2",
     "@ngrx/store": "^13.0.2",
@@ -39,10 +39,10 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.0.4",
-    "@angular/cli": "~13.0.4",
-    "@angular/compiler-cli": "~13.0.3",
-    "@angular/language-service": "^13.0.3",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@babel/core": "^7.10.3",
     "@compodoc/compodoc": "^1.1.16",
     "@types/jasmine": "~3.6.0",
@@ -62,7 +62,7 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.4.4",
+    "typescript": "~4.8.4",
     "webpack": "^5.65.0"
   }
 }
